Move List & Map methods into dedicated files for astbuilder
